refactor: 인수 테스트 공통 사용 기능에 대해서 패키지를 나누어라
