+++ src/test/java/nextstep/subway/acceptance/LineAcceptanceTest.java
@@ -2,6 +2,7 @@ package nextstep.subway.acceptance;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import nextstep.subway.acceptance.common.AcceptanceTest;
 import nextstep.subway.applicaion.dto.LineRequest;
 import nextstep.subway.applicaion.dto.LineResponse;
 import nextstep.subway.applicaion.dto.LineUpdateRequest;
@@ -12,7 +13,7 @@ import org.springframework.http.HttpStatus;
 
 import java.util.List;
 
-import static nextstep.subway.acceptance.RestAssuredTemplate.*;
+import static nextstep.subway.acceptance.common.RestAssuredTemplate.*;
 import static nextstep.subway.acceptance.StationAcceptanceTest.역을_만들다;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
