+++ src/test/java/nextstep/subway/acceptance/common/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.acceptance.common;
 
 
 import io.restassured.RestAssured;
