+++ src/test/java/nextstep/subway/acceptance/common/RestAssuredTemplate.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.acceptance.common;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
